+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 0.1.27 (TBD)
-- Added option to export UV2 data for M2 models as non-standard OBJ property.
+- Added option to export UV2 data for M2 models as non-standard OBJ property (disabled by default).
 - Added texture information (name, dimensions, encoding) to the texture preview window.
+- Added option to export meta data for BLP files (disabled by default).
+- Added option to export meta data for WMO files (disabled by default).
 - Fixed issue that could cause alpha channel to be linked incorrectly in imported material shaders.
 - Fixed issue that preventing pasting into a listfile search under some circumstances.